dreefchat/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pysqlite3-binary
gunicorn
openai
python-dotenv
pydantic
msal
requests
Office365-REST-Python-Client
chromadb
python-multipart
PyPDF2
python-docx
psutil
pyodbc
